+++ src/main/kotlin/nl/sugcube/dirtyarrows/command/DirtyArrowsCommandManager.kt
@@ -22,6 +22,7 @@ open class DirtyArrowsCommandManager(private val plugin: DirtyArrows) : CommandE
     val commands: List<SubCommand<DirtyArrows>> = listOf(
             CommandReload(),
             CommandGive(),
+            CommandAmmo(),
             CommandList(),
             CommandCheck(),
             CommandTeleport(),